requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
langchain==0.3.26
langgraph==0.5.0
langfuse==2.59.7
pydantic-settings==2.10.1
langchain-openai==0.3.27
langchain-community==0.3.26
pdfplumber==0.11.7
PyMuPDF==1.26.1
python-dotenv==1.1.1
faiss-cpu==1.11.0
